refactor: Improve integer type selection
Refactors integer type selection in the OpenAPI to Rust code generator.

Improves the selection of `u32` vs `u64` for integer types based on schema constraints such as maximum value, format, and description keywords, resulting in more efficient and accurate Rust code generation.

Adds a command-line option to override the package description.

+++ crates/algod_client/src/models/avm_value.rs
@@ -17,7 +17,7 @@ use serde::{Deserialize, Serialize};
 pub struct AvmValue {
     /// value type. Value `1` refers to **bytes**, value `2` refers to **uint64**
     #[serde(rename = "type")]
-    pub r#type: u64,
+    pub r#type: u32,
     /// bytes value.
     #[serde(rename = "bytes", skip_serializing_if = "Option::is_none")]
     pub bytes: Option<String>,
@@ -32,7 +32,7 @@ impl AlgorandMsgpack for AvmValue {
 
 impl AvmValue {
     /// Constructor for AvmValue
-    pub fn new(r#type: u64) -> AvmValue {
+    pub fn new(r#type: u32) -> AvmValue {
         AvmValue {
             r#type,
             bytes: None,
